+++ test/rules/unrelated_type_equality_checks.dart
@@ -4,6 +4,8 @@
 
 // test w/ `pub run test -N unrelated_type_equality_checks`
 
+import '../util/fake_fixnum.dart';
+
 void someFunction() {
   var x = '1';
   if (x == 1) print('someFunction'); // LINT
@@ -94,6 +96,30 @@ void someFunction14(DerivedClass4 instance) {
   var other = new DerivedClass5();
 
   if (other == instance) print('someFunction15'); // LINT
+}
+
+void someFunction15() {
+  var x = new Int32();
+
+  if (x == 0) print('someFunction15'); // OK
+}
+
+void someFunction16() {
+  var x = new Int32();
+
+  if (0 == x) print('someFunction16'); // LINT
+}
+
+void someFunction17() {
+  var x = new Int64();
+
+  if (x == 0) print('someFunction17'); // OK
+}
+
+void someFunction18() {
+  var x = new Int64();
+
+  if (0 == x) print('someFunction18'); // LINT
 }
 
 class ClassBase {}