+++ test/util/fake_fixnum.dart
@@ -1,0 +1,5 @@
+library fixnum;
+
+class Int32 {}
+
+class Int64 {}