+++ CHANGELOG.md
@@ -4,6 +4,31 @@
 
 ### Language
 
+* Invocations of noSuchMethod receive default values for optional args.
+  * The following program used to print "No arguments passed", and now prints
+    "First argument is 3".
+
+```dart
+abstract class B {
+  void m([int x = 3]);
+}
+
+class A implements B {
+  noSuchMethod(Invocation i) {
+    if (i.positionalArguments.length == 0) {
+      print("No arguments passed");
+    } else {
+      print("First argument is ${i.positionalArguments[0]}");
+    }
+  }
+}
+
+void main() {
+  A().m();
+}
+```
+
+
 #### Strong Mode
 
 ### Core library changes